+++ src/components/TableView/TableView.tsx
@@ -1,10 +1,10 @@
 import { css, cx } from '@emotion/css';
 import { PanelProps } from '@grafana/data';
 import { Alert, ClickOutsideWrapper, ToolbarButton, ToolbarButtonRow, useTheme2 } from '@grafana/ui';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useRef } from 'react';
 
 import { TEST_IDS } from '../../constants';
-import { useContentPosition, useContentSizes, useLocalStorage, useScrollTo, useTable } from '../../hooks';
+import { useContentPosition, useContentSizes, useSavedState, useScrollTo, useTable } from '../../hooks';
 import { PanelOptions } from '../../types';
 import { Table } from '../Table';
 import { getStyles } from './TableView.styles';
@@ -19,26 +19,13 @@ interface Props extends PanelProps<PanelOptions> {}
  */
 export const TableView: React.FC<Props> = ({ data, id, options, width, height, eventBus }) => {
   /**
-   * useLocalStorage Hook
+   * Current group
    */
-  const { get: getValue, update: saveValue, remove: removeValue } = useLocalStorage(`volkovlabs.variable.panel.${id}`);
-
-  /**
-   * Current Preselected Group
-   */
-  const currentPreselectedGroup = useMemo(() => {
-    if (options.saveSelectedGroup) {
-      const selectedGroup = getValue();
-      return options.groups?.find((group) => group.name === selectedGroup)?.name || options.groups?.[0]?.name;
-    }
-
-    return options.groups?.[0]?.name;
-  }, [getValue, options.groups, options.saveSelectedGroup]);
-
-  /**
-   * Current Levels Group
-   */
-  const [currentGroup, setCurrentGroup] = useState(currentPreselectedGroup);
+  const [currentGroup, setCurrentGroup] = useSavedState<string>({
+    key: `volkovlabs.variable.panel.${options.saveSelectedGroupKey || id}`,
+    initialValue: options.groups?.[0]?.name || '',
+    enabled: options.saveSelectedGroup,
+  });
 
   /**
    * Current Levels
@@ -55,10 +42,9 @@ export const TableView: React.FC<Props> = ({ data, id, options, width, height, e
    */
   useEffect(() => {
     if (!options.groups?.some((group) => group.name === currentGroup)) {
-      removeValue();
-      setCurrentGroup(options.groups?.[0]?.name);
+      setCurrentGroup(options.groups?.[0]?.name || '');
     }
-  }, [currentGroup, id, options.groups, removeValue]);
+  }, [currentGroup, id, options.groups, setCurrentGroup]);
 
   /**
    * Table config
@@ -194,7 +180,6 @@ export const TableView: React.FC<Props> = ({ data, id, options, width, height, e
                     variant={currentGroup === group.name ? 'active' : 'default'}
                     onClick={() => {
                       setCurrentGroup(group.name);
-                      saveValue(group.name);
                     }}
                     data-testid={TEST_IDS.tableView.tab(group.name)}
                     className={styles.toolbarButton}