+++ src/module.test.ts
@@ -31,6 +31,7 @@ describe('plugin', () => {
     addRadio: jest.fn().mockImplementation(() => builder),
     addSelect: jest.fn().mockImplementation(() => builder),
     addSliderInput: jest.fn().mockImplementation(() => builder),
+    addTextInput: jest.fn().mockImplementation(() => builder),
   };
 
   it('Should be instance of PanelPlugin', () => {
@@ -51,12 +52,14 @@ describe('plugin', () => {
     expect(builder.addRadio).toHaveBeenCalled();
     expect(builder.addSelect).toHaveBeenCalled();
     expect(builder.addSliderInput).toHaveBeenCalled();
+    expect(builder.addTextInput).toHaveBeenCalled();
   });
 
   describe('Input Visibility', () => {
     beforeEach(() => {
       builder.addFieldNamePicker.mockClear();
       builder.addSelect.mockClear();
+      builder.addTextInput.mockClear();
     });
 
     /**
@@ -140,6 +143,17 @@ describe('plugin', () => {
       plugin['optionsSupplier'](builder);
 
       expect(shownOptionsPaths).toEqual(expect.arrayContaining(['padding']));
+    });
+
+    it('Should show Unique Cross Key Input if table view', () => {
+      const shownOptionsPaths: string[] = [];
+
+      builder.addTextInput.mockImplementation(
+        addInputImplementation({ displayMode: DisplayMode.TABLE, saveSelectedGroup: true }, shownOptionsPaths)
+      );
+      plugin['optionsSupplier'](builder);
+
+      expect(shownOptionsPaths).toEqual(expect.arrayContaining([]));
     });
   });